refactor product components
