+++ src/components/buttons/cartFloatButton.jsx
@@ -6,7 +6,7 @@ import { Link, useNavigate } from 'react-router-dom'
 import { useSelector } from 'react-redux'
 
 import useScrollPosition from '../../hooks/useScrollPosition'
-import ProductCartList from '../productCartList'
+import { ProductCartList } from '../products'
 import Button from './button'
 import { selectCartItems } from '../../redux/features/carts/cartSlice'
 import { selectCurrentStep } from '../../redux/features/stepper/stepperSlice'
