+++ src/components/products/index.js
@@ -1,0 +1,21 @@
+import ProductCarousel from './productCarousel'
+import ProductCartList from './productCartList'
+import ProductCheckout from './productCheckout'
+import ProductCheckoutList from './productCheckoutList'
+import ProductCheckoutListButton from './productCheckoutListButton'
+import ProductDesc from './productDesc'
+import ProductItem from './productItem'
+import ProductItemListing from './productItemListing'
+import ProductScrollView from './productScrollView'
+
+export {
+  ProductCarousel,
+  ProductCartList,
+  ProductCheckout,
+  ProductCheckoutList,
+  ProductCheckoutListButton,
+  ProductDesc,
+  ProductItem,
+  ProductItemListing,
+  ProductScrollView
+}
