+++ src/components/products/productDesc.jsx
@@ -13,7 +13,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faMinus, faPlus } from '@fortawesome/free-solid-svg-icons'
 import { useState } from 'react'
 
-import { Button } from './buttons'
+import { Button } from '../buttons'
 
 function ProductDesc({
   name,
