+++ src/pages/home.jsx
@@ -7,7 +7,7 @@ import homepage from '../assets/images/homepage.png'
 import feature2 from '../assets/images/features2.png'
 import { Button, LinkButton } from '../components/buttons'
 import EmailSignUp from '../components/emailSignUp'
-import ProductItemListing from '../components/productItemListing'
+import { ProductItemListing } from '../components/products'
 import {
   selectStatus,
   selectProducts,
