+++ src/pages/products/category.jsx
@@ -5,7 +5,7 @@ import { useParams } from 'react-router-dom'
 
 import clubImg from '../../assets/images/features3.png'
 import EmailField from '../../components/emailField'
-import ProductItemListing from '../../components/productItemListing'
+import { ProductItemListing } from '../../components/products'
 import { Button } from '../../components/buttons'
 import {
   selectProductsCategory,
