+++ src/pages/products/productListing.jsx
@@ -5,7 +5,7 @@ import { Helmet } from 'react-helmet-async'
 import { Button } from '../../components/buttons'
 import clubImg from '../../assets/images/features3.png'
 import EmailField from '../../components/emailField'
-import ProductItemListing from '../../components/productItemListing'
+import { ProductItemListing } from '../../components/products'
 import {
   getProducts,
   selectProducts,
