+++ src/pages/user/wishlist.jsx
@@ -1,11 +1,12 @@
-import ProductItemListing from '../../components/productItemListing'
+import { useDispatch, useSelector } from 'react-redux'
+import { useEffect } from 'react'
+
+import { ProductItemListing } from '../../components/products'
 import {
   getWishlist,
   selectWishlistList,
   selectWishlistStatus
 } from '../../redux/features/wishlist/wishlistSlice'
-import { useDispatch, useSelector } from 'react-redux'
-import { useEffect } from 'react'
 import { selectUserUid } from '../../redux/features/userSlice'
 import PagePreloader from '../../components/preloader/pagePreloader'
